requirements.txt
Flask
msoffcrypto-tool
openpyxl
pandas
